+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 851df32 - v148
+| Author: fjrodafo - Mon, 20 Jan 2025 21:15:02 +0000
+| 
 * Commit: ff5a848 - v147
 | Author: fjrodafo - Sun, 19 Jan 2025 23:29:38 +0000
 |